fix: skip alert divs for endpoint descriptions
